master: multi_checkpoint.cpp: fix conversion warning when using boost vectors.

+++ test_more/cppad_for_tmb/multi_checkpoint.cpp
@@ -83,7 +83,8 @@ bool multi_checkpoint(void)
 		v         = f.Forward(0, x);
 		//
 		// this assigment has false sharing; i.e., will case cache resets
-		y[thread] = v[0];
+		// (conversion seems to avoid boost vector conversion warning)
+		y[size_t(thread)] = v[0];
 	}
 
 	// check the results
